+++ angel_serialize_generator/test/models/has_map.g.dart
@@ -29,6 +29,11 @@ class HasMap implements _HasMap {
     return hashObjects([value]);
   }
 
+  @override
+  String toString() {
+    return "HasMap(value=$value)";
+  }
+
   Map<String, dynamic> toJson() {
     return HasMapSerializer.toMap(this);
   }
@@ -38,7 +43,29 @@ class HasMap implements _HasMap {
 // SerializerGenerator
 // **************************************************************************
 
-abstract class HasMapSerializer {
+const HasMapSerializer hasMapSerializer = const HasMapSerializer();
+
+class HasMapEncoder extends Converter<HasMap, Map> {
+  const HasMapEncoder();
+
+  @override
+  Map convert(HasMap model) => HasMapSerializer.toMap(model);
+}
+
+class HasMapDecoder extends Converter<Map, HasMap> {
+  const HasMapDecoder();
+
+  @override
+  HasMap convert(Map map) => HasMapSerializer.fromMap(map);
+}
+
+class HasMapSerializer extends Codec<HasMap, Map> {
+  const HasMapSerializer();
+
+  @override
+  get encoder => const HasMapEncoder();
+  @override
+  get decoder => const HasMapDecoder();
   static HasMap fromMap(Map map) {
     if (map['value'] == null) {
       throw new FormatException("Missing required field 'value' on HasMap.");